Move the 'LibraryNode' class to a separate file
